jobs/requirements.txt
azure-storage-blob==12.14.1
requests==2.31.0